#75 added new storage field for different placement for items

+++ storage-module/src/main/resources/db/migration/V1__characters_create.sql
@@ -1,10 +1,10 @@
 create table if not exists characters
 (
     character_name text primary key,
-    account_name   text     not null,
-    experience     integer  not null,
-    x              integer  not null,
-    y              integer  not null,
-    angle          smallint not null,
+    account_name   text             not null,
+    experience     integer          not null,
+    x              double precision not null,
+    y              double precision not null,
+    angle          smallint         not null,
     unique (character_name, account_name)
 );
